gthemedicon: Fix a -Wsign-conversion warning for some string handling
The result of the subtraction is guaranteed to be non-negative by the
code above, so the cast is safe.


Helps: #3405

+++ gio/gthemedicon.c
@@ -295,7 +295,7 @@ g_themed_icon_update_names (GThemedIcon *themed)
               gchar *tmp = last;
               gchar *fallback;
 
-              last = g_strndup (last, dashp - last);
+              last = g_strndup (last, (size_t) (dashp - last));
               if (is_symbolic)
                 {
                   g_free (tmp);
